06-is-it-down-monitor/requirements.txt
python-dotenv
fastapi
jetpack-io>=0.5.1.dev202112011638382467
requests
uvicorn
